requirements.txt
Flask==3.1.1
numpy==2.0.2
pillow==11.3.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
torch==2.8.0
transformers==4.55.2
